LPS-147594 Split files to routes folder
