+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/routes/Outlets/ActivationOutlet.js
@@ -11,10 +11,10 @@
 
 import {useEffect} from 'react';
 import {Outlet, useNavigate, useOutletContext} from 'react-router-dom';
-import {PAGE_TYPES, PRODUCT_TYPES} from '../../utils/constants';
-import ActivationKeys from './ActivationKeys';
-import DXP from './DXP';
-import DXPCloud from './DXPCloud';
+import {PAGE_TYPES, PRODUCT_TYPES} from '../../../../utils/constants';
+import ActivationKeys from '../../ActivationKeys';
+import DXP from '../../DXP';
+import DXPCloud from '../../DXPCloud';
 
 const ACTIVATION_ROOT_ROUTER = 'activation';
