+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/routes/Outlets/ProductsOutlet.js
@@ -11,7 +11,7 @@
 
 import {useMemo} from 'react';
 import {useOutletContext, useParams} from 'react-router-dom';
-import {PAGE_TYPES, PRODUCT_TYPES} from '../../utils/constants';
+import {PAGE_TYPES, PRODUCT_TYPES} from '../../../../utils/constants';
 
 const ProductsOutlet = () => {
 	const {productId} = useParams();
